+++ benchmarks/server/raw/http.func.ts
@@ -1,6 +1,8 @@
-import { useFunc, useHttp } from '@faasjs/core'
+import { Func, Http } from '@faasjs/core'
 
-export const func = useFunc(() => {
-  useHttp()
-  return async () => 'Hello'
+export const func = new Func({
+  plugins: [new Http()],
+  async handler() {
+    return 'Hello'
+  },
 })